requirements.txt
Django==4.1.10
django-crispy-forms==2.0
django-extensions==3.2.3
django-role-permissions==3.2.0
crispy-bootstrap5==0.7

